refactor: First pass at simplification of the crate's code
In particular, this first pass renames various functions, modules, structs to more logical names.
